requirements.txt
Flask==3.0.0
scapy==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
PyYAML==6.0.1
pytest==7.4.3
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2